refactored auth fetches to useFetch
