+++ frontend/src/pages/LoginPage.tsx
@@ -1,13 +1,13 @@
 import { FormEvent, useState } from "react";
+import {useNavigate} from "react-router-dom";
+import {useFetch} from "../hooks/useFetch.ts";
 
-
-type Props = {
-    onLogin: (event: FormEvent, userName: string, password: string) => void
-}
-export default function LoginPage({onLogin}: Props) {
+export default function LoginPage() {
 
     const [userName, setUserName] = useState("");
     const [password, setPassword] = useState("");
+    const navigate = useNavigate();
+    const login = useFetch((state) => state.login);
 
     function handleUserNameInput(event: FormEvent<HTMLInputElement>) {
         setUserName(event.currentTarget.value);
@@ -20,7 +20,7 @@ export default function LoginPage({onLogin}: Props) {
     return (
         <>
             <h2>Login</h2>
-            <form onSubmit={(event) => onLogin(event, userName, password)}>
+            <form onSubmit={(event) => login(event, userName, password, navigate)}>
                 <label>
                     Username:
                     <input value={userName} onChange={handleUserNameInput} type="text" name="userName" placeholder="username"/>
